requirements.txt
uagents
uagents-core
requests
openai
python-dotenv

